refactor: Template for resource & data source & schema files
