+++ internal/provider/connection/schema.go
@@ -1,24 +1,14 @@
 package connection
 
 import (
-	"context"
-
 	"terraform-provider-hookdeck/internal/validators"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 )
-
-func (r *connectionResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
-	resp.Schema = schema.Schema{
-		MarkdownDescription: "Connection Resource",
-		Attributes:          schemaAttributes(),
-	}
-}
 
 func schemaAttributes() map[string]schema.Attribute {
 	filterRulePropertySchema := schema.SingleNestedAttribute{
