+++ internal/provider/destination/resource.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	hookdeckClient "github.com/hookdeck/hookdeck-go-sdk/client"
 )
 
@@ -29,6 +30,14 @@ type destinationResource struct {
 // Metadata returns the resource type name.
 func (r *destinationResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
 	resp.TypeName = req.ProviderTypeName + "_destination"
+}
+
+// Schema returns the resource schema.
+func (r *destinationResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		MarkdownDescription: "Destination Resource",
+		Attributes:          schemaAttributes(),
+	}
 }
 
 // Configure adds the provider configured client to the resource.
