+++ internal/provider/transformation/resource.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	hookdeck "github.com/hookdeck/hookdeck-go-sdk"
 	hookdeckClient "github.com/hookdeck/hookdeck-go-sdk/client"
 )
@@ -31,6 +32,14 @@ type transformationResource struct {
 // Metadata returns the resource type name.
 func (r *transformationResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
 	resp.TypeName = req.ProviderTypeName + "_transformation"
+}
+
+// Schema returns the resource schema.
+func (r *transformationResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		Description: "Transformation Resource",
+		Attributes:  schemaAttributes(),
+	}
 }
 
 // Configure adds the provider configured client to the resource.
